backend/requirements.txt
Flask~=3.0.3
flask_cors
PyYAML~=6.0.1
orange3